1. Added drag-and-drop functionality for dragging a tab from inactive to active session. 2. Temporarily fixed the bug where an 'about:' page cannot be dragged between active tabs. 3. Both of the above have been done by modifying Application::loadUrl() by adding a MainWindow parameter, following Andrea's suggestion.

+++ src/sessionmanagement/sessiontabdata.cpp
@@ -142,8 +142,12 @@ void SessionTabData::updateThumbnail(int width, int height)
         m_thumbnail = WebSnap::renderTabPreview(*m_webTab.data()->page(), width , height);
     }
     else
-    { 
+    {
         m_thumbnail = WebSnap::renderTabPreview(*m_webTab.data()->page(), prevWidth , prevHeight);
+    }
+    if (m_thumbnail.isNull())
+    {
+        return;
     }
     QString path = WebSnap::imagePathFromUrl(m_url);
     QFile::remove(path);
